FULLY WORKING AND CLEAN AND NORME --> TODO %llf

+++ libft/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: omulder <omulder@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/20 20:30:44 by omulder        #+#    #+#                */
-/*   Updated: 2019/02/20 20:31:25 by omulder       ########   odam.nl         */
+/*   Updated: 2019/02/28 14:58:19 by omulder       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		ft_abs(int n)
+long long		ft_abs(long long n)
 {
 	if (n < 0)
 		return (-n);
